+++ src/client/client_serializer.h
@@ -22,9 +22,9 @@ public:
 
     std::vector<char> serialize_shooting();
 
-    std::vector<char> serialize_throw_explosive_grenede(const int8_t& time);
+    std::vector<char> serialize_throw_explosive_grenede(const int32_t& time);
 
-    std::vector<char> serialize_throw_smoke_grenede(const int8_t& time);
+    std::vector<char> serialize_throw_smoke_grenede(const int32_t& time);
 
     std::vector<char> serialize_call_air_strike();
